Consolidate the IsMockCreation() method to use semantic model.
